requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
google-generativeai==0.7.2
python-dotenv==1.0.1
numpy==1.26.4
faiss-cpu==1.8.0.post1 